Expanded/changed display text, image size
I shrank the pokemon image sprites down a bit. I deleted the title text
and moved the species name up to take its place. Below the image I now
display the Pokemon's typing. One message for monotype, and a second
for dualtypes (n/a is displayed if there is no second type).