Add music settings object, barebones search menu
I create a music settings object to handle the program's music settings.
I would like to implement this in the original file as well if this goes
smoothly. I created a barebones menu page that will become the pokedex
search page. I added a return button to the pokemon display page that
currently is not properly functional