created pokemon object, updated readme
I created a pokemon object to track the basic attributes that my
display uses so far. I also changed the pokemon display function
in order to take account of these changes. I updated the readme
with the first mention of my pokeAPI spin-off along with some
todo's for the project.