Created the species display in pokemon_search
I have created a cycle that displays pokemon (currently 9 at a time)
based on their order in the national pokedex onto the screen. I have
added a next button that will always produce the next 9 pokemon as
well. I also created a Pokedex class and object to keep track of
elements pertaining to this execution.