change font sizes, adjusted UI, added back button
Made a bunch of small changes. Adjusted the positioning and order of
buttons on the pokemon search page so it now looks much better. I
added a back button to go back in the pokedex list offsets. Changed
the colours of a couple buttons. Made it so that the offsets don't
reset every time you return to the page.