Renamed GameScreen vars, created text function
I renamed the GameScreen dimensions to height and width, since I
figured that having "screen" in the name was redundant now. I also
created a function to set render text and blit it in one fell swoop.
Seems to work fine but not all text is converted yet and it needs a
bit of testing.