Added index search to pokeAPI, renamed files
The two Pokedexes have been renamed. The now-primary one is the newer
one which takes advantage of pokeAPI. I renamed it pokedex_2.0.
The original was renamed pokedex_1.0 to reflect the progression
between them. I added the ability to search the pokedex list in 2.0
by inputting an index number. I also made a pretty substantial revamp
to the readme.