Simplify pokemon search list, updated readme
Now, instead of calling the api every time the user changes the page,
I've just called the entire list at the beginning of the program,
storing it in a variable called pokeAPI_master_list. Also made it so
that backspaces in the dex number search also updates the list, and
reduced the character limit on the search bar to 4, since the number of
pokemon only gets up to 4 digits and I don't think it will reach 5 for
a very long time.